job-alert-automator/requirements.txt
requests>=2.31.0
schedule>=1.2.0
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth>=2.28.0